Address PR feedback for 26.1 port
- Remove unused ItemStack import from CreateWorldScreenMixin
- Guard empty targetItemId in ClientHuntState.handleWin to avoid null
- Cache display name alongside ItemStack in HuntHudRenderer
- Add fast-path flag and try-catch around bindComponents in ItemPool
- Compare MouseButtonEvent x/y as doubles for sub-pixel accuracy

+++ src/main/java/com/spawnhunt/screen/SpawnHuntScreen.java
@@ -341,8 +341,8 @@ public class SpawnHuntScreen extends Screen {
 
     @Override
     public boolean mouseClicked(MouseButtonEvent event, boolean selected) {
-        if (event.button() == 0 && (int) event.x() >= historyLinkX && (int) event.x() <= historyLinkX + historyLinkW
-                && (int) event.y() >= historyLinkY && (int) event.y() <= historyLinkY + historyLinkH) {
+        if (event.button() == 0 && event.x() >= historyLinkX && event.x() <= historyLinkX + historyLinkW
+                && event.y() >= historyLinkY && event.y() <= historyLinkY + historyLinkH) {
             showHistory = !showHistory;
             return true;
         }
